+++ R/utils_downloads.R
@@ -267,6 +267,24 @@ gsc_api_cache <-
 
     gsc_message(verbose, "Downloading from ", url, "\n")
 
+    # Testing purposes only
+    # Mock the behavior of offline
+    test <- getOption("giscoR_test_offline", NULL)
+
+    if (isTRUE(test)) {
+      gsc_message(
+        TRUE,
+        "\nurl \n ",
+        url,
+        " not reachable.\n\nPlease download manually. ",
+        "If you think this is a bug please consider opening an issue on ",
+        "https://github.com/ropengov/giscoR/issues"
+      )
+      message("Returning `NULL`")
+      return(NULL)
+    }
+
+
     err_dwnload <- suppressWarnings(try(
       download.file(url, file.local, quiet = isFALSE(verbose), mode = "wb"),
       silent = TRUE