+++ tests/testthat/test-gisco_get_postalcodes.R
@@ -5,7 +5,6 @@ test_that("Error on postal codes", {
 test_that("Postal codes online", {
   skip_on_cran()
   skip_if_gisco_offline()
-  skip("Postal Codes may be failing")
 
   expect_message(gisco_get_postalcodes(
     country = "Malta",
@@ -16,4 +15,14 @@ test_that("Postal codes online", {
   li <- expect_silent(gisco_get_postalcodes(country = "Malta"))
   expect_length(unique(li$CNTR_ID), 1)
   expect_identical(as.character(unique(li$CNTR_ID)), "MT")
+})
+
+test_that("offline", {
+  options(giscoR_test_offline = TRUE)
+  expect_message(
+    n <- gisco_get_postalcodes(update_cache = TRUE, country = "Malta"),
+    "not reachable"
+  )
+  expect_null(n)
+  options(giscoR_test_offline = FALSE)
 })